# FE/aiPhaseII/src/api/ai二期.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ai二期",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/virtualOrders/import/studentSubsidy": {
      "post": {
        "summary": "导入学生补贴表",
        "deprecated": false,
        "description": "上传Excel文件，导入学生补贴信息并自动生成虚拟任务",
        "operationId": "import_student_subsidy_api_virtualOrders_import_studentSubsidy_post",
        "tags": [
          "虚拟订单服务"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "\n        返回数据结构:\n        ```json\n        {\n            \"code\": 200,\n            \"message\": \"导入成功\",\n            \"data\": {\n                \"import_batch\": \"批次号\",\n                \"total_students\": 10,\n                \"total_subsidy\": 2000.00,\n                \"generated_tasks\": 45\n            }\n        }\n        ```\n        ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/virtualOrders/import/customerService": {
      "post": {
        "summary": "导入专用客服",
        "deprecated": false,
        "description": "上传Excel文件，导入专用客服信息",
        "operationId": "import_customer_service_api_virtualOrders_import_customerService_post",
        "tags": [
          "虚拟订单服务"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "\n        返回数据结构:\n        ```json\n        {\n            \"code\": 200,\n            \"message\": \"导入成功\",\n            \"data\": {\n                \"total_imported\": 5,\n                \"failed_count\": 0\n            }\n        }\n        ```\n        ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/api/virtualOrders/studentPools": {
      "get": {
        "summary": "获取学生补贴池列表",
        "deprecated": false,
        "description": "分页获取学生补贴池信息",
        "operationId": "get_student_pools_api_virtualOrders_studentPools_get",
        "tags": [
          "虚拟订单服务"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "页码",
            "required": false,
            "schema": {
              "title": "Page",
              "minimum": 1,
              "type": "integer",
              "description": "页码",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "每页数量",
            "required": false,
            "schema": {
              "title": "Size",
              "maximum": 100,
              "minimum": 1,
              "type": "integer",
              "description": "每页数量",
              "default": 20
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "状态过滤",
            "required": false,
            "schema": {
              "title": "Status",
              "type": "string",
              "description": "状态过滤"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "\n        返回数据结构:\n        ```json\n        {\n            \"code\": 200,\n            \"message\": \"获取成功\",\n            \"data\": {\n                \"items\": [...],\n                \"total\": 100,\n                \"page\": 1,\n                \"size\": 20\n            }\n        }\n        ```\n        ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/virtualOrders/reallocate/{studentId}": {
      "post": {
        "summary": "重新分配学生任务",
        "deprecated": false,
        "description": "手动触发学生剩余金额的任务重新分配",
        "operationId": "reallocate_student_tasks_api_virtualOrders_reallocate__student_id__post",
        "tags": [
          "虚拟订单服务"
        ],
        "parameters": [
          {
            "name": "student_id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "title": "Student Id",
              "type": "integer"
            }
          },
          {
            "name": "studentId",
            "in": "path",
            "description": "",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "\n        返回数据结构:\n        ```json\n        {\n            \"code\": 200,\n            \"message\": \"重新分配成功\",\n            \"data\": {\n                \"student_id\": 123,\n                \"remaining_amount\": 150.00,\n                \"new_tasks_count\": 8\n            }\n        }\n        ```\n        ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/virtualOrders/customerService": {
      "get": {
        "summary": "获取虚拟客服列表",
        "deprecated": false,
        "description": "分页获取虚拟订单专用客服列表",
        "operationId": "get_virtual_customer_services_api_virtualOrders_customerService_get",
        "tags": [
          "虚拟订单服务"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "页码",
            "required": false,
            "schema": {
              "title": "Page",
              "minimum": 1,
              "type": "integer",
              "description": "页码",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "每页数量",
            "required": false,
            "schema": {
              "title": "Size",
              "maximum": 100,
              "minimum": 1,
              "type": "integer",
              "description": "每页数量",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "\n        返回数据结构:\n        ```json\n        {\n            \"code\": 200,\n            \"message\": \"获取成功\",\n            \"data\": {\n                \"items\": [...],\n                \"total\": 10,\n                \"page\": 1,\n                \"size\": 20\n            }\n        }\n        ```\n        ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "post": {
        "summary": "创建虚拟客服",
        "deprecated": false,
        "description": "创建虚拟订单专用客服账号",
        "operationId": "create_virtual_customer_service_api_virtualOrders_customerService_post",
        "tags": [
          "虚拟订单服务"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VirtualCustomerServiceCreate"
              },
              "example": ""
            }
          }
        },
        "responses": {
          "200": {
            "description": "\n        返回数据结构:\n        ```json\n        {\n            \"code\": 200,\n            \"message\": \"创建成功\",\n            \"data\": {\n                \"id\": 123,\n                \"user_id\": 456,\n                \"name\": \"客服姓名\",\n                \"account\": \"客服账号\",\n                \"level\": \"6\",\n                \"status\": \"active\",\n                \"initial_password\": \"123456\"\n            }\n        }\n        ```\n        ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/virtualOrders/customerService/{csId}": {
      "put": {
        "summary": "更新虚拟客服信息",
        "deprecated": false,
        "description": "更新虚拟订单专用客服的基本信息",
        "operationId": "update_virtual_customer_service_api_virtualOrders_customerService__cs_id__put",
        "tags": [
          "虚拟订单服务"
        ],
        "parameters": [
          {
            "name": "cs_id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "title": "Cs Id",
              "type": "integer"
            }
          },
          {
            "name": "csId",
            "in": "path",
            "description": "",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VirtualCustomerServiceUpdate"
              },
              "example": ""
            }
          }
        },
        "responses": {
          "200": {
            "description": "\n        返回数据结构:\n        ```json\n        {\n            \"code\": 200,\n            \"message\": \"更新成功\",\n            \"data\": {\n                \"id\": 456,\n                \"name\": \"新姓名\",\n                \"account\": \"客服账号\",\n                \"status\": \"active\",\n                \"updated_fields\": [\"name\", \"status\"]\n            }\n        }\n        ```\n        ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "delete": {
        "summary": "删除虚拟客服",
        "deprecated": false,
        "description": "删除虚拟订单专用客服（软删除）",
        "operationId": "delete_virtual_customer_service_api_virtualOrders_customerService__cs_id__delete",
        "tags": [
          "虚拟订单服务"
        ],
        "parameters": [
          {
            "name": "cs_id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "title": "Cs Id",
              "type": "integer"
            }
          },
          {
            "name": "csId",
            "in": "path",
            "description": "",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "\n        返回数据结构:\n        ```json\n        {\n            \"code\": 200,\n            \"message\": \"删除成功\",\n            \"data\": {\n                \"id\": 456,\n                \"name\": \"客服姓名\",\n                \"account\": \"客服账号\",\n                \"deleted\": true\n            }\n        }\n        ```\n        ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/virtualOrders/studentIncome/export": {
      "get": {
        "summary": "导出学生收入数据",
        "deprecated": false,
        "description": "导出学生收入统计数据为Excel文件，用于后续补贴金额设置",
        "operationId": "export_student_income_api_virtualOrders_studentIncome_export_get",
        "tags": [
          "虚拟订单服务"
        ],
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "description": "开始日期",
            "required": false,
            "schema": {
              "title": "Start Date",
              "type": "string",
              "description": "开始日期",
              "format": "date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "description": "结束日期",
            "required": false,
            "schema": {
              "title": "End Date",
              "type": "string",
              "description": "结束日期",
              "format": "date"
            }
          },
          {
            "name": "student_ids",
            "in": "query",
            "description": "学生ID列表，多个ID用逗号分隔",
            "required": false,
            "schema": {
              "title": "Student Ids",
              "type": "array",
              "items": {
                "type": "integer"
              },
              "description": "学生ID列表，多个ID用逗号分隔"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "\n        返回Excel文件下载:\n        - Content-Type: application/vnd.openxmlformats-officedocument.spreadsheetml.sheet\n        - 包含学生收入统计和使用说明两个工作表\n        - 补贴金额列供用户填写后重新导入\n        ",
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "headers": {}
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/virtualOrders/studentIncome/summary": {
      "get": {
        "summary": "获取学生收入汇总",
        "deprecated": false,
        "description": "获取学生收入的汇总统计信息",
        "operationId": "get_student_income_summary_api_virtualOrders_studentIncome_summary_get",
        "tags": [
          "虚拟订单服务"
        ],
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "description": "开始日期",
            "required": false,
            "schema": {
              "title": "Start Date",
              "type": "string",
              "description": "开始日期",
              "format": "date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "description": "结束日期",
            "required": false,
            "schema": {
              "title": "End Date",
              "type": "string",
              "description": "结束日期",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "\n        返回数据结构:\n        ```json\n        {\n            \"code\": 200,\n            \"message\": \"获取成功\",\n            \"data\": {\n                \"total_students\": 100,\n                \"total_tasks\": 500,\n                \"total_amount\": 25000.00,\n                \"completed_tasks\": 450,\n                \"completed_amount\": 22500.00,\n                \"completion_rate\": 90.0,\n                \"export_time\": \"2023-12-01 10:30:00\"\n            }\n        }\n        ```\n        ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
...
      "VirtualCustomerServiceCreate": {
        "title": "VirtualCustomerServiceCreate",
        "required": [
          "name",
          "account"
        ],
        "type": "object",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string",
            "description": "客服姓名"
          },
          "account": {
            "title": "Account",
            "type": "string",
            "description": "客服账号"
          },
          "initialPassword": {
            "title": "Initial Password",
            "type": "string",
            "description": "初始密码",
            "default": "123456"
          }
        },
        "description": "创建虚拟客服请求"
      },
...
      "VirtualCustomerServiceUpdate": {
        "title": "VirtualCustomerServiceUpdate",
        "type": "object",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string",
            "description": "客服姓名"
          },
          "status": {
            "title": "Status",
            "type": "string",
            "description": "状态：active-活跃, inactive-停用"
          }
        },
        "description": "更新虚拟客服请求"
      },
...
    },
    "securitySchemes": {}
  },
  "servers": [],
  "security": []
}